[EPMCDLAB-1384]: Fixed removing remote kernels from notebook image

+++ infrastructure-provisioning/src/general/lib/os/fab.py
@@ -472,11 +472,24 @@ def replace_multi_symbols(string, symbol, symbol_cut=False):
         traceback.print_exc(file=sys.stdout)
 
 
-def update_pyopenssl_lib():
+def update_pyopenssl_lib(os_user):
+    if not exists('/home/{}/.ensure_dir/pyopenssl_updated'.format(os_user)):
+        try:
+            if exists('/usr/bin/pip3'):
+                sudo('pip3 install -U pyopenssl')
+            sudo('pip2 install -U pyopenssl')
+            sudo('touch /home/{}/.ensure_dir/pyopenssl_updated'.format(os_user))
+        except:
+            sys.exit(1)
+
+
+def find_cluster_kernels():
     try:
-        print("Updating pyOpenssl lib")
-        if exists('/usr/bin/pip3'):
-            sudo('pip3 install -U pyopenssl')
-        sudo('pip2 install -U pyopenssl')
+        with settings(sudo_user='root'):
+            de = [i for i in sudo('find /opt/ -maxdepth 1 -name "*-de-*" -type d | rev | '
+                                  'cut -f 1 -d "/" | rev | xargs -r').split(' ') if i != '']
+            des =  [i for i in sudo('find /opt/ -maxdepth 2 -name "*-des-*" -type d | rev | '
+                                    'cut -f 1,2 -d "/" | rev | xargs -r').split(' ') if i != '']
+        return (de, des)
     except:
         sys.exit(1)
